+++ migrations/000005_add_path_groups.down.sql
@@ -1,0 +1,2 @@
+DROP TABLE IF EXISTS path_group_members;
+DROP TABLE IF EXISTS path_groups;